% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{acceptance_rejection}
\alias{acceptance_rejection}
\title{Acceptance and rejection method for generating pseudo-random numbers}
\usage{
acceptance_rejection(n = 1L, pdf, ...)
}
\arguments{
\item{n}{Number of generations to be generated.}

\item{pdf}{Probability density function defined on the interval \eqn{0 < x < 1}.}

\item{...}{Probability density function parameters.}
}
\description{
Acceptance and rejection method for generating pseudo-random numbers from any
distribution in the interval \eqn{0 < x < 1}.
}
\details{
The accept and reject method implemented by this function is
intended to generate pseudo random observations of random variables in the
interval \eqn{0 < x < 1}. Therefore, the passed probability density function pdf must
be a function with supported defined in this range.
}
\examples{
# Example (Simplex)
acceptance_rejection(n = 1e3L, pdf = pdf_simplex, mu = 0.9, sigma2 = 0.9) |>
hist(prob = TRUE)
x <- seq(0, 1, length.out = 200)
y <- pdf_simplex(x = x, mu = 0.9, sigma2 = 0.9)
lines(x, y, type = "l")
# Example (Beta)
acceptance_rejection(n = 1e3, pdf = pdf_beta, mu = 0.2, phi = 0.2) |>
hist(prob = TRUE)
x <- seq(0, 1, length.out = 200)
y <- pdf_beta(x = x, mu = 0.2, phi = 0.2)
lines(x, y, type = "l")
# Example (Unit Gamma)
acceptance_rejection(n = 1e3, pdf = pdf_gamma_u, mu = 0.7, tau = 1.34) |>
hist(prob = TRUE)
x <- seq(0, 1, length.out = 200)
y <- pdf_gamma_u(x = x, mu = 0.7, tau = 1.34)
lines(x, y, type = "l")
}
