% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pdf_beta}
\alias{pdf_beta}
\title{Beta distribution (reparameterized)}
\usage{
pdf_beta(x, mu, phi)
}
\arguments{
\item{x}{Value in the domain of the function, such that \eqn{0 < x < 1}.}

\item{mu}{Lease parameter, such that \eqn{0 < \mu < 1}.}

\item{psi}{Dispersion parameter, such that \eqn{\phi^2 > 0}.}
}
\description{
Probability density function of a random variable with beta distribution.
}
\examples{
integrate(f = pdf_beta, lower = 0, upper = 1, mu = 0.5, phi = 1.2)
}
