% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pdf_gamma_u}
\alias{pdf_gamma_u}
\title{Unit gamma distribution}
\usage{
pdf_gamma_u(x, mu, tau)
}
\arguments{
\item{x}{Value in the domain of the function, such that \eqn{0 < x < 1}.}

\item{mu}{Lease parameter, such that \eqn{0 < \mu < 1}.}

\item{tau}{Dispersion parameter, such that \eqn{\tau^2 > 0}.}
}
\description{
Probability density function of a random variable with unit gamma distribution.
}
\examples{
integrate(f = pdf_gamma_u, lower = 0, upper = 1, mu = 0.5, tau = 1.2)
}
