% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{pdf_simplex}
\alias{pdf_simplex}
\title{Simplex distribution}
\usage{
pdf_simplex(x, mu, sigma2)
}
\arguments{
\item{x}{Value in the domain of the function, such that \eqn{0 < x < 1}.}

\item{mu}{Lease parameter, such that \eqn{0 < \mu < 1}.}

\item{sigma2}{Dispersion parameter, such that \eqn{\sigma^2 > 0}.}
}
\description{
Probability density function of a random variable with simplex distribution.
}
\examples{
integrate(f = pdf_simplex, lower = 0, upper = 1, mu = 0.5, sigma2 = 1.2)
}
